Navbar product qa (#1656)
* Weglot initialization script after hydration

* Declares Weglot variable

* Make devhub trigger medium weight

* Remove NavbarView

* Aligns caret triggers with dropdown arrow

* Hide desktop logo on mobile

* Style the crap out of the network selector

* More styling

* Revert "Remove NavbarView"

This reverts commit f437b6e40604119b7763418089ad966cbe837d32.

* Serves icons from public folder

* Fix svg paths

* Removes NavBarView, moves hiding logic into hook

---------
