+++ src/components/Header/AlgoSearch/searchButton.module.css
@@ -1,9 +1,10 @@
 .default {
   display: flex;
   align-items: center;
+  padding: var(--space-2x);
+
   border: 1px solid rgba(255, 255, 255, 0);
   border-radius: var(--space-1x);
-  padding: var(--space-2x);
   overflow: hidden;
   white-space: nowrap;
   width: fit-content;
@@ -31,10 +32,11 @@
   display: none;
 }
 
-@media screen and (min-width: 37.75em) {
-  .container {
+@media screen and (min-width: 996px) {
+  .default {
     max-width: 160px; /* Initial width */
     transition: max-width 0.5s ease, border 0.5s ease; /* Transitions for width and border */
+    padding: 11px var(--space-2x);
   }
 
   .default:hover {
@@ -50,7 +52,11 @@
   }
 }
 
-@media screen and (min-width: 50em) {
+@media screen and (min-width: 1080px) {
+  .default {
+    padding: 5px var(--space-2x);
+  }
+
   .default .shortcut {
     display: flex;
     padding: var(--space-1x) var(--space-2x);
