--- src/components/Header/Nav/ProductNavigation/Desktop/ArrowIcon.tsx
@@ -1,12 +1,0 @@
-export const ArrowIcon = () => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M16 12.25H7" stroke="#FAFAFB" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-    <path
-      d="M12.5 8.25L16.25 12L12.5 15.75"
-      stroke="#FAFAFB"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-)
