+++ src/components/Header/Nav/ProductNavigation/Desktop/productNavigation.module.css
@@ -5,9 +5,20 @@
   z-index: 1;
 }
 
+.logo {
+  display: none;
+}
+
 @media screen and (min-width: 768px) {
   .root {
     display: flex;
+  }
+
+  .logo {
+    display: flex;
+    align-items: center;
+    padding-block: var(--space-1x);
+    display: block;
   }
 }
 
@@ -20,13 +31,6 @@
   align-items: center;
   list-style: none;
   margin: 0;
-}
-
-.logo {
-  display: flex;
-  align-items: center;
-  padding-block: var(--space-1x);
-  display: block;
 }
 
 .button {
@@ -59,21 +63,12 @@
   max-width: 90dvw;
 }
 
-.content[data-motion="from-start"] {
-  animation-name: enterFromLeft;
-}
-.content[data-motion="from-end"] {
-  animation-name: enterFromRight;
-}
-.content[data-motion="to-start"] {
-  animation-name: exitToLeft;
-}
-.content[data-motion="to-end"] {
-  animation-name: exitToRight;
-}
-
 .content li {
   list-style: none;
+}
+
+.productTrigger span {
+  font-weight: var(--font-weight-medium);
 }
 
 .indicator {
@@ -98,7 +93,6 @@
   display: flex;
   justify-content: center;
   top: 100%;
-  /* perspective: 2000px; */
 }
 
 .navigationViewport {
@@ -121,57 +115,14 @@
 }
 
 .arrow {
-  position: relative;
+  position: absolute;
   top: 70%;
   background-color: var(--white);
   width: 10px;
   height: 10px;
+  right: var(--space-1x);
   transform: rotate(45deg);
   border-top-left-radius: 2px;
-}
-
-@keyframes enterFromRight {
-  from {
-    opacity: 0;
-    transform: translateX(200px);
-  }
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-@keyframes enterFromLeft {
-  from {
-    opacity: 0;
-    transform: translateX(-200px);
-  }
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-@keyframes exitToRight {
-  from {
-    opacity: 1;
-    transform: translateX(0);
-  }
-  to {
-    opacity: 0;
-    transform: translateX(200px);
-  }
-}
-
-@keyframes exitToLeft {
-  from {
-    opacity: 1;
-    transform: translateX(0);
-  }
-  to {
-    opacity: 0;
-    transform: translateX(-200px);
-  }
 }
 
 @keyframes scaleIn {
