+++ src/components/Header/Nav/ProductNavigation/Mobile/Category.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { ProductItem } from "../../config"
-import { clsx, getImageUrl } from "../../utils"
+import { clsx } from "../../utils"
 import { CaretRightIcon } from "./CaretRightIcon"
 import styles from "./category.module.css"
 import { SubProducts } from "./ProductNavigation"
@@ -14,7 +14,7 @@ const Item = React.forwardRef<HTMLAnchorElement, ListItemProps>(
   ({ item: { label, icon, href, subProducts }, onProductClick }, forwardedRef) => {
     const itemComponent = (
       <>
-        {icon && <img height={24} width={24} src={getImageUrl(`/${icon}-navbar-icon.svg`)} />}
+        {icon && <img height={24} width={24} src={`/assets/icons/${icon}-navbar-icon.svg`} />}
         <span style={{ flex: 1, textAlign: "start" }} className="text-300">
           {label}
         </span>
