+++ src/components/Header/Nav/ProductNavigation/Mobile/ResourcesIcon.tsx
