+++ src/components/Header/Nav/navBar.module.css
@@ -48,49 +48,92 @@
   display: none;
 }
 
+:global(.wg-drop.country-selector .wgcurrent.wg-li) {
+  border: none;
+  background: none;
+  display: flex;
+  height: var(--space-10x);
+  gap: var(--space-2x);
+  padding: var(--space-2x) var(--space-2x) var(--space-2x) 0;
+  line-height: var(--space-4x);
+  justify-content: center;
+  align-items: center;
+  color: var(--blue-700);
+  border: var(--border-width-primary) solid transparent;
+  border-radius: var(--border-radius-primary);
+  transition: background-color 300ms ease-out, border 300ms ease-out, border-radius 300ms ease-out;
+}
+
+:global(.wg-drop.country-selector .wgcurrent.wg-li):hover {
+  border: var(--border-width-primary) solid var(--blue-200);
+  background: var(--background-color-background-primary);
+}
+
+:global(.wg-drop.country-selector) {
+  background: none !important;
+}
+
+:global(.wg-drop.country-selector .wg-current a) {
+  padding-left: 0 !important;
+  padding-right: 24px !important;
+  color: var(--color-text-primary) !important;
+}
+
+:global(#weglot-listbox) {
+  border-radius: var(--border-radius-primary);
+  color: var(--color-text-primary) !important;
+}
+
+:global(.country-selector a) {
+  color: var(--color-text-primary) !important;
+  padding-right: 24px !important;
+}
+:global(.country-selector .wgcurrent:hover a),
+:global(.country-selector :not(.wgcurrent) a:hover) {
+  color: var(--color-text-link) !important;
+}
+
+:global(.wg-drop.country-selector .wgcurrent:after) {
+  background-image: url("/assets/icons/caret.svg") !important;
+}
+
+:global(.wg-drop.country-selector .wgcurrent:hover:after) {
+  background-image: url("/assets/icons/caret-blue.svg") !important;
+}
+
+:global(.wg-drop.country-selector.closed .wgcurrent:after) {
+  transform: rotate(0deg) !important;
+}
+
+:global(.wg-drop.country-selector.open .wgcurrent:after) {
+  transform: rotate(-180deg) !important;
+}
+
 @media screen and (min-width: 768px) {
-  :is(.button, .iconButton) {
+  :is(.button) {
     display: flex;
     height: var(--space-10x);
     gap: var(--space-2x);
-    padding: 0 var(--space-4x);
+    padding: var(--space-2x);
     line-height: var(--space-4x);
     justify-content: center;
     align-items: center;
     color: var(--blue-700);
     border: var(--border-width-primary) solid transparent;
-    transition: opacity 300ms ease-out, background-color 300ms ease-out, border 300ms ease-out,
-      border-radius 300ms ease-out;
-    opacity: 1;
+    transition: background-color 300ms ease-out, border 300ms ease-out, border-radius 300ms ease-out;
   }
 
-  :is(.button, .iconButton):hover {
+  :is(.button):hover {
     border-radius: var(--border-radius-primary);
     border: var(--border-width-primary) solid var(--blue-200);
     background: var(--background-color-background-primary);
-    opacity: 0.9;
-  }
-
-  .iconButton {
-    padding: var(--space-2x);
-    width: var(--space-5x);
-    height: var(--space-5x);
-    min-width: var(--space-5x);
-    min-height: var(--space-5x);
-  }
-
-  .iconButton img {
-    width: var(--space-5x);
-    height: var(--space-5x);
-    min-width: var(--space-5x);
-    min-height: var(--space-5x);
   }
 
   :is(.button) :is(svg, img) {
-    width: var(--space-4x);
-    height: var(--space-4x);
-    min-width: var(--space-4x);
-    min-height: var(--space-4x);
+    width: var(--space-5x);
+    height: var(--space-5x);
+    min-width: var(--space-5x);
+    min-height: var(--space-5x);
     fill: var(--color-text-primary);
   }
