+++ src/components/Header/Nav/utils.ts
@@ -1,51 +1,3 @@
-export const IMGIX_URL = "https://smartcontract.imgix.net"
-
-type DPR = 1 | 2 | 3 | 4 | 5
-
-type SearchParams = {
-  border?: string
-  "border-radius"?: number
-  txt?: string
-  w?: number
-  h?: number
-  dpr?: DPR
-  bg?: string
-  auto?: string
-  pad?: number
-  txt64?: string
-  "txt-font"?: string
-  "txt-color"?: string
-  "txt-align"?: string
-  "txt-size"?: string
-  "txt-fit"?: string
-  fit?: "fillmax"
-  fill?: "solid"
-  trim?: string
-}
-
-export function buildUrl(baseUrl: string, params: SearchParams = {}) {
-  const url = new URL(baseUrl)
-
-  Object.entries(params).forEach(([key, val]) => {
-    if (val) {
-      url.searchParams.set(key, val.toString())
-    }
-  })
-
-  return url.toString()
-}
-
-export function getImageUrl(imagePath: string, params?: SearchParams) {
-  return buildUrl(`${IMGIX_URL}${imagePath}`, {
-    auto: "compress,format",
-    ...params,
-  })
-}
-
-export function getIconUrl(iconName: string) {
-  return getImageUrl(`/icons/${iconName}.svg`)
-}
-
 /**
  * This is a copy from "shared" library.
  * A tiny utility for constructing className strings conditionally.
