+++ src/components/Header/NavBar.tsx
@@ -43,7 +43,7 @@ export const NavBar = ({ path, showSearch = true }: { path: string; showSearch?:
                 full_name: false,
                 with_name: true,
                 is_dropdown: true,
-                with_flags: false,
+                with_flags: true,
                 flag_type: "circle",
               },
               location: {
